Fix open GUI with key binding
